Computing colocation quotient (CLQ) on permutated data

Randomly shuffles the cell type labels within each tissue section and computes permutated CLQs between every cell type pair, creating a null distribution that is compared with the observed CLQ to compute a permutation test p-value.

Usage

CLQ_permutation_test(CLQ_obs, CLQ_perm)

Arguments

CLQ_obs

Observed CLQs computed using compute_CLQ_observed().

CLQ_perm

Permutated CLQs computed using compute_CLQ_permutated().

Value

A data frame containing the nominal and Benjamini-Hochberg-adjusted CLQ permutation test p-values.

Details

The permutation test p-value of tissue section \(l\) is defined as

$$p_{b\rightarrow a}^{(l)} = \sum_{k=1}^{\mbox{iter.num}}I(CLQ_{b\rightarrow a}^{(l), k} > CLQ_{b\rightarrow a}^{(l), obs}),$$

where \(CLQ_{b\rightarrow a}^{(l), k}\) is the permutated CLQ computed using compute_CLQ_permutated() on the randomly shuffled cell type labels, and \(CLQ_{b\rightarrow a}^{(l), obs}\) is the observed CLQ computed using compute_CLQ_observed() on the true cell type labels.

References

Bouchard, G. et al. A quantitative spatial cell-cell colocalizations framework enabling comparisons between in vitro assembloids and pathological specimens. Nat Commun 16, 1392 (2025). https://doi.org/10.1038/s41467-024-55129-6