Prepare input data CTN dendrogram

ggdendro package is required for the function.

Usage

prepare_dendro_data(cl, h, k = NULL)

Arguments

cl

An object of class "hclust" which describes the tree produced by the clustering process. Output of the jaccard_dist_hclust() function.

h

Numeric scalar or vector with heights where the tree should be cut. stats::cutree() for details. At least one of k or h must be specified, k overrides h if both are given.

k

An integer scalar or vector with the desired number of groups. See stats::cutree() for details.

Value

A list with components (See ggdendro::dendro_data() for details):

segments

Line segment data

labels

Label data

Details

Internal function for preparing input data to draw_CTN_dendro() function

References

Atrebas. Dendrograms in R, a lightweight approach. https://atrebas.github.io/post/2019-06-08-lightweight-dendrograms/